% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAltModeTrips.R
\name{applyHurdleTripModel}
\alias{applyHurdleTripModel}
\title{Apply a hurdle trip model}
\usage{
applyHurdleTripModel(Data_df, Model_ls)
}
\arguments{
\item{Data_df}{a data frame containing the independent variables needed to
calculate trips}

\item{Model_ls}{a list created by the estimateHurdleTripModel which contains
'Count' and 'Zero' model components.}
}
\value{
a numeric vector of the number of trips for each household in the
same order as Data_df
}
\description{
\code{applyHurdleTripModel} applies a hurdle model to calculate the number
of trips for each household.
}
\details{
This function takes a list of strings representing an estimated hurdle model
and applies it to a data frame that contains data of independent variables
for household records. it returns a vector of modeled trips corresponding to
the rows of the data frame.
}
