% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHouseholdDVMT.R
\name{CalculateHouseholdDVMT}
\alias{CalculateHouseholdDVMT}
\title{Calculate the average household DVMT, 95th percentile household DVMT, and
maximum household DVMT.}
\usage{
CalculateHouseholdDVMT(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateHouseholdDVMT} calculate the average household DVMT, 95th
percentile household DVMT, and maximum household DVMT.
}
\details{
This function calculates the average household DVMT, 95th percentile
household DVMT, and maximum household DVMT as a function of the household
characteristics and the characteristics of the area where the household
resides.
}
