% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHouseholdDVMT.R
\docType{data}
\name{DvmtModel_ls}
\alias{DvmtModel_ls}
\title{Daily vehicle miles traveled (DVMT) models}
\format{A list having 'Metro' and 'NonMetro' components. Each component has
the following components:
Pow: factor to untransform the results of the average DVMT model;
Ave: the formula for the average DVMT model;
Pctl95: the formula for the 95th percentile DVMT model;
Max: the formula for the maximum DVMT model.}
\source{
GreenSTEP version 3.6 model.
}
\usage{
DvmtModel_ls
}
\description{
A list of components describing the average daily vehicle miles traveled
models for metropolitan and non-metropolitan areas. The models are linear
models which predict average DVMT, 95th percentile DVMT, and maximum DVMT.
The models are linear regression models. The average DVMT model predicts
average DVMT as a power transform. The 'Pow' component contains the power
term for untransforming the model results. The 95th percentile and maximum
DVMT models predict the households 95th percentile and maximum daily travel
as a function of the household's average DVMT.
}
\keyword{datasets}
